v1 23/12 by Ngon Phi

+++ Sources/application/views/admin/list_user_admin_view.php
@@ -32,7 +32,7 @@
                         <td align="center">
                             <a class="btn btn-default" href="<?php echo base_url();?>index.php/admin/edit_user/<?php echo $row['id_tk'];?>"><em class="fa fa-pencil"></em></a>
                             <br>
-                            <a class="btn btn-danger" href="<?php echo base_url();?>index.php/admin/delete_user/<?php echo $row['id_tk'];?>"><em class="fa fa-trash"></em></a>
+                            <a class="btn btn-danger" onclick="return confirm('Bạn có muốn xóa không');" href="<?php echo base_url();?>index.php/admin/delete_user/<?php echo $row['id_tk'];?>"><em class="fa fa-trash"></em></a>
                         </td>
                         <td><?php echo $row['id_tk'];?></td>
                         <td><?php echo $row['ten_dang_nhap'];?></td>
